+++ packages/bitcore-wallet-client/lib/key.js
@@ -15,7 +15,15 @@ var Constants = Common.Constants;
 var Utils = Common.Utils;
 var Credentials = require('./credentials');
 
-var FIELDS = [
+
+function Key() {
+  this.version = '1.0.0';
+  this.use145forBCH = true;
+  this.use48forMultisig = true;
+  this.compliantDerivation = true;
+};
+
+Key.FIELDS = [
   'xPrivKey',             // obsolte
   'xPrivKeyEncrypted',   // obsolte
   'mnemonic',
@@ -28,12 +36,6 @@ var FIELDS = [
   'version',
 ];
 
-function Key() {
-  this.version = '1.0.0';
-  this.use145forBCH = true;
-  this.use48forMultisig = true;
-  this.compliantDerivation = true;
-};
 
 const wordsForLang = {
   'en': Mnemonic.Words.ENGLISH,
@@ -121,7 +123,7 @@ Key.fromObj = function(obj) {
     throw 'Bad Key version';
   }
 
-  _.each(FIELDS, function(k) {
+  _.each(Key.FIELDS, function(k) {
     x[k] = obj[k];
   });
 
@@ -130,48 +132,12 @@ Key.fromObj = function(obj) {
 };
 
 
-Key.fromOld = function(x) {
-  $.shouldBeObject(x);
-
-  // TODO RO clients
-  if (!_.isUndefined(x.version) || (!x.xPrivKey && !x.xPrivKeyEncrypted)) {
-    throw 'Could not recognize old version';
-  }
-
-  var k = new Key();
-  _.each(FIELDS, (i) => {
-    if (!_.isUndefined(x[i])) {
-      k[i] = x[i];
-    }
-  });
-
-  // ALL old credentials have multi wallets in 44'.
-  k.use48forMultisig = false;
-
-  let oldFIELDS = [
-  ];
-
-  var c = new Credentials();
-  _.each(Credentials.FIELDS, function(i) {
-    c[i] = x[i];
-  });
-  if (c.externalSource) {
-    throw new Error('External Wallets are no longer supported');
-  }
-  c.coin = c.coin || 'btc';
-  c.addressType = c.addressType || Constants.SCRIPT_TYPES.P2SH;
-  c.account = c.account || 0;
-  c.rootPath = c.getRootPath();
-  return {key: k, credentials: c};
-};
-
-
 
 Key.prototype.toObj = function() {
   var self = this;
 
   var x = {};
-  _.each(FIELDS, function(k) {
+  _.each(Key.FIELDS, function(k) {
     x[k] = self[k];
   });
   return x;
